Rename to thank-you; update BEM; update tests
