+++ app/components/thank-you-container.js
@@ -1,0 +1,9 @@
+import Ember from 'ember';
+
+const {
+  Component
+} = Ember;
+
+export default Component.extend({
+  classNames: ['thank-you-container']
+});
