+++ app/router.js
@@ -57,7 +57,7 @@ AppRouter.map(function() {
       this.route('task', { path: '/:number' });
     });
     this.route('donate');
-    this.route('thankyou');
+    this.route('thank-you');
   });
 
   this.route('projects', {
