+++ app/styles/app.scss
@@ -83,7 +83,7 @@
 @import "templates/about";
 @import "templates/oauth-stripe";
 @import "templates/project/index";
-@import "templates/project/thankyou";
+@import "templates/project/thank-you";
 @import "templates/project/settings/contributors";
 @import "templates/start/expertise";
 @import "templates/start/hello";
