--- app/styles/templates/project/thankyou.scss
@@ -1,27 +1,0 @@
-.thankyou-container {
-  .icon, .header, .text, .project-link {
-    width: 100%;
-    text-align: center;
-  }
-
-  .header, .text {
-    margin-bottom: 20px;
-  }
-
-  .header {
-    font-size: 1.5em;
-    margin-top: 0px;
-  }
-
-  .icon {
-    img {
-      display: inline-block;
-    }
-    margin-bottom: 0px;
-  }
-
-  .text {
-    color: $light-text;
-    font-size: 14px;
-  }
-}
