+++ app/templates/components/thank-you-container.hbs
@@ -1,0 +1,10 @@
+<div class="thank-you-container__project-icon" data-test-selector="project icon"></div>
+<h3>Thank you!</h3>
+<div class="thank-you-container__message">
+  <p data-test-selector="thank you message">
+    From all the volunteers on the {{project.title}} team.
+  </p>
+  <p>
+    {{link-to "Back to project" "project" project.organization.slug project.slug}}
+  </p>
+</div>
