--- app/templates/components/thankyou-container.hbs
@@ -1,4 +1,0 @@
-<div class="icon"><img src={{project.iconLargeUrl}} alt={{projectIconAltText}} data-test-selector="project icon" /></div>
-<h3 class="header">Thank you!</h3>
-<div class="text" data-test-selector="thank you message">From all the volunteers on the {{project.title}} team.</div>
-<div class="project-link">{{#link-to "project" project}}Back to project{{/link-to}}</div>
