+++ app/templates/project/thank-you.hbs
@@ -1,0 +1,2 @@
+{{project-header project=model}}
+{{thank-you-container project=model}}
