--- app/templates/project/thankyou.hbs
@@ -1,2 +1,0 @@
-{{project-header project=model}}
-{{thankyou-container project=model}}
