+++ tests/acceptance/project-thank-you-test.js
@@ -1,0 +1,49 @@
+import { test } from 'qunit';
+import moduleForAcceptance from 'code-corps-ember/tests/helpers/module-for-acceptance';
+
+import { authenticateSession } from 'code-corps-ember/tests/helpers/ember-simple-auth';
+import createProjectWithSluggedRoute from 'code-corps-ember/tests/helpers/mirage/create-project-with-slugged-route';
+import projectThankYouPage from '../pages/project/thank-you';
+
+moduleForAcceptance('Acceptance | Project - thank-you');
+
+test('It requires authentication', function(assert) {
+  assert.expect(1);
+
+  let project = createProjectWithSluggedRoute();
+  let { organization } = project;
+
+  projectThankYouPage.visit({
+    organization: organization.slug,
+    project: project.slug
+  });
+
+  andThen(() => {
+    assert.equal(currentRouteName(), 'login');
+  });
+});
+
+test('It directs you to the project when done', function(assert) {
+  assert.expect(2);
+
+  let user = server.create('user');
+  authenticateSession(this.application, { 'user_id': user.id });
+
+  let project = createProjectWithSluggedRoute();
+  let { organization } = project;
+
+  projectThankYouPage.visit({
+    organization: organization.slug,
+    project: project.slug
+  });
+
+  andThen(() => {
+    assert.ok(projectThankYouPage.thankYouContainer.isVisible, 'The thank you container component renders.');
+
+    projectThankYouPage.thankYouContainer.clickLink();
+  });
+
+  andThen(() => {
+    assert.equal(currentRouteName(), 'project.index', 'Link leads to project page.');
+  });
+});
