+++ tests/pages/components/thank-you-container.js
@@ -1,14 +1,16 @@
 import {
-  attribute,
+  clickable,
   text
 } from 'ember-cli-page-object';
 
 export default {
+  scope: '.thank-you-container',
+
   icon: {
-    scope: '[data-test-selector="project icon"]',
-    alt: attribute('alt'),
-    src: attribute('src')
+    scope: '[data-test-selector="project icon"]'
   },
+
+  clickLink: clickable('a'),
 
   thankYouText: text('[data-test-selector="thank you message"]')
 };
